TASK-2 and TASK-3 completed

+++ backend/grails-app/domain/backend/Stock.groovy
@@ -4,14 +4,13 @@ import java.time.LocalDateTime
 
 class Stock {
 
-    BigInteger price
+    Double price
     LocalDateTime priceDate
 
     static belongsTo = [company: Company]
 
     static constraints = {
         price nullable: false, blank: false
-
         priceDate nullable: false, blank: false
     }
 }
